+++ semana-15/Aula 43/template-apis-rest/src/index.ts
@@ -1,7 +1,7 @@
 //importando express com Request e Response e cors
 import express, {Request, Response} from 'express';
 import cors from 'cors';
-
+import exercicio1 from './exercicio1'
 //extra: importando configuração de rede do node
 import { AddressInfo } from "net";
 //iniciando a aplicação web com express
@@ -11,59 +11,7 @@ const app = express();
 app.use(express.json());
 app.use(cors());
 
-type user = {
-    id: number,
-    name: string,
-    email: string,
-    type: string,
-    age: number
-}
 
-
-let users: user[] = [
-    {
-        id: 1,
-        name: "Alice",
-        email: "alice@email.com",
-        type: "ADMIN",
-        age: 12
-    },
-    {
-        id: 2,
-        name: "Bob",
-        email: "bob@email.com",
-        type: "NORMAL",
-        age: 36
-    },
-    {
-        id: 3,
-        name: "Coragem",
-        email: "coragem@email.com",
-        type: "NORMAL",
-        age: 21
-    },
-    {
-        id: 4,
-        name: "Dory",
-        email: "dory@email.com",
-        type: "NORMAL",
-        age: 17
-    },
-    {
-        id: 5,
-        name: "Elsa",
-        email: "elsa@email.com",
-        type: "ADMIN",
-        age: 17
-    },
-    {
-        id: 6,
-        name: "Fred",
-        email: "fred@email.com",
-        type: "ADMIN",
-        age: 60
-    }
-]
